% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{char_stats_tbl}
\alias{char_stats_tbl}
\title{char_stats_tbl}
\format{A data frame with a unique item for each categorical variable:
   1. State Initial
   2. Timestamp, when the summary stats were created
   3. Min date range used to query data
   4. Max date range used to query data
   5. Variable Name
   6. Category
   7. Category Percent
   8. Sample Size}
\usage{
char_stats_tbl
}
\description{
Statistics for categorical variables. Categorical variable statistics from all states live in this data frame.
}
\keyword{datasets}

