% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_classes_match}
\alias{check_classes_match}
\title{check_classes_match}
\usage{
check_classes_match(df)
}
\arguments{
\item{df}{- A data.frame}
}
\description{
Check if the variable classes in the new state match the expected class.
If the classes do not match then an error message will be displayed. This function is applied
within the pull_data function. The variable classes will be compared to the expected variable classes in the data dictionary.
}

