% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_var_range}
\alias{check_var_range}
\title{check_var_range}
\usage{
check_var_range(df1, df2 = NA, df3 = NA, df4 = NA, df5 = NA, df6 = NA,
  df7 = NA)
}
\arguments{
\item{df'x'}{- any data frame that contains variables you would like to check the range of.}
}
\description{
Check the range of a variable pulled from the NRDB. Is it within
the acceptable range? Variable ranges are compared to ranges set in the data dictionary.
}

