% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data.R
\name{compare_data}
\alias{compare_data}
\title{compare_data}
\usage{
compare_data(st_initials = NULL, char = FALSE, numeric = FALSE)
}
\arguments{
\item{st_initials}{- a vector of state initials.  input should look like c("id", "ks", "nc", .....)}

\item{char}{- Do you want to compare the categorical variables. TRUE or FALSE}

\item{numeric}{- Do you want to compare the numeric variables. TRUE or FALSE}
}
\description{
Compare variable statistics between multiple states.
This function displays the visualizations that show the differences between states.
}

