% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tbls.R
\name{create_char_tbl}
\alias{create_char_tbl}
\title{create_char_tbl}
\usage{
create_char_tbl(df, st_initials, min_date, max_date)
}
\arguments{
\item{df}{- a data.frame of the categorical variables}

\item{st_initials}{- Initials of the state you are generating stats for}
}
\description{
Calculate the percent that each category makes up within each categorical variable. Store them in a data frame.
This function is applied in the generate_new_stats function.
}

