% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_var_descriptions_tbl.R
\name{create_data_descriptions}
\alias{create_data_descriptions}
\title{create_data_descriptions}
\usage{
create_data_descriptions()
}
\description{
Make a table containing a description of each variable. The table will contain the majority of information
 in the data dictionary. It is applied in the form_data_dictionary function.
}

