% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tbls.R
\name{create_density_tbl}
\alias{create_density_tbl}
\title{create_density_tbl}
\usage{
create_density_tbl(df, st_initials, min_date, max_date)
}
\arguments{
\item{df}{- A data.frame containing numeric variables to validate}

\item{st_initials}{- Initials of the state you are generating stats for}
}
\description{
Calculate a density function for each variable and store them in a data frame.
This function is applied in the generate_new_stats function.
}

