% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tbls.R
\name{create_num_tbl}
\alias{create_num_tbl}
\title{create_num_tbl}
\usage{
create_num_tbl(df, st_initials, min_date, max_date)
}
\arguments{
\item{df}{- A data.frame containing numeric variables to validate}

\item{st_initials}{- Initials of the state you are generating stats for}

\item{min_date}{- the minimum date that the data is pulled from}

\item{max_date}{- the max date that the data is pulled from}
}
\description{
Create a table containing statistics on each numeric variable and store them in a data frame.
This function is applied in the generate_new_stats function.
}

