% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_var_range_tbl.R
\name{create_var_range_tbl}
\alias{create_var_range_tbl}
\title{create_var_range_tbl}
\usage{
create_var_range_tbl()
}
\description{
Create a table to set an acceptable range for each variable. Used in the form_data_dictionary function
}

