% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{density_tbl}
\alias{density_tbl}
\title{density_tbl}
\format{A data frame containing information required to
make density plots for each numeric variable
   1. State Initial
   2. Timestamp, when the summary stats were created
   3. Min date range used to query data
   4. Max date range used to query data
   5. Variable Name
   6. pdf x-values
   7. pdf y-values}
\usage{
density_tbl
}
\description{
Probability Density Function x and y values for numeric variables in each
state live in this data frame.
}
\keyword{datasets}

