% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dimension}
\alias{dimension}
\title{dimension}
\usage{
dimension(df)
}
\arguments{
\item{df}{- A data.frame}
}
\description{
Calculates the number of rows and columns, and displays the values to the user
when new data is pulled. *For the users reference
}

