% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{form_data_dictionary}
\alias{form_data_dictionary}
\title{form_data_dictionary}
\usage{
form_data_dictionary()
}
\description{
create a data dictionary of the data used to generate scores. This
dictionary contains name, class, parent table in the NRDB, table it is stored in locally,
description, min acceptable value and max acceptable value
}

