% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_new_stats.R
\name{generate_new_stats}
\alias{generate_new_stats}
\title{generate_new_stats}
\usage{
generate_new_stats(st_initials = "AA", min_date = "1900-01-01",
  max_date = Sys.Date(), save = FALSE, pull = TRUE)
}
\description{
Generates new statistics by running the functions
 create_char_tbl, create_density_tbl, create_num_tbl across all variables
}

