% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{num_stats_tbl}
\alias{num_stats_tbl}
\title{num_stats_tbl}
\format{A data frame containing descriptive statistics for each numeric variable
   1. State Initial
   2. Timestamp, when the summary stats were created
   3. Min date range used to query data
   4. Max date range used to query data
   5. Variable Name
   6. Statistic
   7. Statistic Value}
\usage{
num_stats_tbl
}
\description{
Statistics for numeric variables. Numeric variable statistics from all states live in this data frame.
}
\keyword{datasets}

