% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{pull data}
\usage{
pull_data(min_date = "1900-01-01", max_date = Sys.Date())
}
\arguments{
\item{min_date}{- the minimum date to filter the query by. of the form 'year-month-day'}

\item{max_date}{- the maximum date to filter the query by. of the form 'year-month-day'}
}
\description{
Pull all data fields that are used for generating scores.
This includes the statistical modules and business rules.
The output includes 8 data.frames created from tables in the NRDB
}

