% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarity_measures}
\alias{similarity_measures}
\title{similarity_measures}
\usage{
similarity_measures(pdf1_x, pdf1_y, pdf2_x, pdf2_y)
}
\arguments{
\item{pdf1_x}{- x values from a pdf}

\item{pdf1_y}{- y values from a pdf}

\item{pdf2_x}{- x values from a pdf}

\item{pdf2_y}{- y values from a pdf}
}
\description{
Estimate the bhattacharyya coefficient and shared area
between two probability density functions.
}

