% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{variable_classes}
\alias{variable_classes}
\title{variable_classes}
\format{A data.frame of expected variable classes}
\usage{
variable_classes
}
\description{
data frame of expected variable classes for each variable used in scoring
}
\keyword{datasets}

